Accomodate fraction of a minute in a expiry time of token

+++ app/shared/configs/settings.py
@@ -22,8 +22,8 @@ class Settings(BaseSettings):
     # JWT Secret Key
     JWT_SECRET: str = config("JWT_SECRET", "649fb93ef34e4fdf4187709c84d643dd61ce730d91856418fdcf563f895ea40f")
     JWT_ALGORITHM: str = config("ACCESS_TOKEN_ALGORITHM", "HS256")
-    ACCESS_TOKEN_EXPIRE_MINUTES: int = config("ACCESS_TOKEN_EXPIRE_MINUTES", default=10, cast=int)
-    REFRESH_TOKEN_EXPIRE_MINUTES: int = config("REFRESH_TOKEN_EXPIRE_MINUTES", default=1440, cast=int)
+    ACCESS_TOKEN_EXPIRE_MINUTES: float = config("ACCESS_TOKEN_EXPIRE_MINUTES", default=10, cast=float)
+    REFRESH_TOKEN_EXPIRE_MINUTES: float = config("REFRESH_TOKEN_EXPIRE_MINUTES", default=1440, cast=float)
 
     # App Secret Key
     SECRET_KEY: str = config("SECRET_KEY", default="8deadce9449770680910741063cd0a3fe0acb62a8978661f421bbcbb66dc41f1")
